// src/enums/Error.ts
export enum EnumErrors {
	ERROR_UNKNOWN = "Une erreur est survenue.",
	ERROR_PERMISSION = "Vous n'avez pas les droits pour effectuer cette action.",
	SESSION_EXPIRED = "La session est expir\u00E9e",
	NO_USER_LOGGED = "Aucun utilisateur connect\u00E9",
	INVALID_TOKEN = "Token de session invalide",
	ERROR_USER_EXIST = "Cet utilisateur existe d\u00E9j\u00E0"
}